Big update.
Data is now stored in a snapshot of the database lists for each data
type using the Serializable interface. Equipment implements comparable
and checks against serial numbers to determine similar equipment items.
testDatabase is a tester function that shows the basic functionality of
the project. The package temporaryDelete was a trash can for some of the
previously uploaded files that should be deleted.